익셉션 핸들러 제작 및 wanted 컨트롤러 적용 (#49)
* 가짜 WantedService제작 후 기능 구현함

* 제작한 가짜 WantedService와 실제 WantedService합침

* WantedDao 정리함

* WantedDao 위치 수정함

* Wanted 매퍼 수정함, WantedsDao 수정함

* Wanted의 Service, Controller 수정함

* Wanted 테스트용 가상 컨트롤러, 서비스 제작완료함

* wanted Mapper수정함

* likes Mapper 수정함

* like mapper, service, controller 구현함

* like mapper, service, controller 수정함함

* subcrives mapper, dao, service 구현함

* Code(Region, skill, career, position) Dao, Mapper, Service 작업함

* 테스트파일 삭제함

* 단위 테스트 제작

* 마스터 병합함

* apply mapper연결 밑 service 구현함

* subcribes mapper 연결 및 service 구현, Dto 구현함

* WantedController 구현 중, LikesService 추가함

* Apply, Wanted 서비스, 컨트롤러 제작함

* Dto static 수정함

* 마스터 pull 함

* wanted 컨트롤러 작업함

* Wanted 컨트롤러 제작함

* Wanted, Company, like Serviece 예외처리 제외 구현함

* wanted, company 컨트롤러 구현 완료

* 익셉션 핸들러 제작함

+++ src/main/java/site/metacoding/miniproject2/service/LikesService.java
@@ -16,6 +16,7 @@ import site.metacoding.miniproject2.dto.LikesRespDto.LikeFindByIdRespDto;
 import site.metacoding.miniproject2.dto.LikesRespDto.LikeInsertRespDto;
 import site.metacoding.miniproject2.dto.SessionUsers;
 import site.metacoding.miniproject2.dto.WantedsRespDto.WantedLisLikestRespDto;
+import site.metacoding.miniproject2.handler.MyApiException;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -38,9 +39,12 @@ public class LikesService {
     }
 
     public LikeInsertRespDto insert(LikesInsertReqDto likesInsertReqDto) {
+        SessionUsers principal = (SessionUsers) session.getAttribute("principal");
+        likesInsertReqDto.setUserId(principal.getId());
+
         // 중복 체크
-        if (likesDao.findByLike(likesInsertReqDto) == null) {
-            throw new RuntimeException("해당 공고에 이미 좋아요를 했습니다.");
+        if (likesDao.findByLike(likesInsertReqDto) != null) {
+            throw new MyApiException("해당 공고에 이미 좋아요를 했습니다.");
         }
 
         likesDao.insert(likesInsertReqDto);
@@ -54,16 +58,12 @@ public class LikesService {
     @Transactional(rollbackFor = RuntimeException.class)
     public void delete(LikesInsertReqDto likesInsertReqDto) {
         SessionUsers principal = (SessionUsers) session.getAttribute("principal");
-        // if (likesDao.findByLike(likesInsertReqDto) != null) {
-        // throw new RuntimeException("잘못된 접근입니다. 1");
-        // }
-        // if (likesDao.findByLike(likesInsertReqDto).getUserId() != principal.getId())
-        // {
-        // throw new RuntimeException("잘못된 접근입니다.");
-        // }
-        // if (likesDao.findByLike(likesInsertReqDto) == null) {
-        // throw new RuntimeException("이미 취소되었습니다.");
-        // }
+        likesInsertReqDto.setUserId(principal.getId());
+
+        // 중복 체크
+        if (likesDao.findByLike(likesInsertReqDto) == null) {
+            throw new MyApiException("좋아요를 취소할 수 없습니다.");
+        }
         likesDao.delete(likesInsertReqDto);
     }
     /* 승현 작업 종료 */
